version 1.1.1

+++ memmgr.cpp
@@ -104,7 +104,7 @@ void  operator  delete(void  *pointer) throw()
 
 	if(pointer)
 	{
-		if(!MEMMGR_is_pointer_registered((char *)pointer-sizeof(PointerPreHeapInfo)))
+		if(!MEMMGR_is_pointer_registered((uintptr_t)((char *)pointer-sizeof(PointerPreHeapInfo))))
 		{
 			LOG_ERROR("(%s:%i): allocated_pointer NOT REGISTERED OR POSSIBLE MEMORY CORRUPTION?!?!",source_file,  source_line);
 		}
@@ -160,7 +160,7 @@ void  operator  delete[](void  *pointer) throw()
 	if(pointer!=NULL)
 	{
 
-		if(!MEMMGR_is_pointer_registered((char *)pointer-sizeof(PointerPreHeapInfo)))
+		if(!MEMMGR_is_pointer_registered((uintptr_t)((char *)pointer-sizeof(PointerPreHeapInfo))))
 		{
 			LOG_ERROR("(%s:%i): allocated_pointer NOT REGISTERED WITH MALLOC OR NEW!",source_file,  source_line);
 		}
